重新梳理代码结构
