+++ config.json
@@ -9,7 +9,7 @@
     "backup_save_path": "/home/even/mysql_backup/",
 
     # 版本填写 5.7 或 8.0
-    "DB_VERSION": "5.7",
+    "DB_VERSION": "8.0",
     "DB_HOST": "127.0.0.1",
     "DB_PORT": 3306,
     "DB_USER": "root",
@@ -20,7 +20,7 @@
     "remote_backup": true,
     "REMOTE_HOST": "192.168.56.15",
     "REMOTE_PORT": 22,
-    "REMOTE_USER": "even",
+    "REMOTE_USER": "kali",
     "REMOTE_PASSWORD": "2333",
     "REMOTE_PATH": "/home/even/test"
 }
