+++ controller/controller.go
@@ -2,7 +2,6 @@ package controller
 
 import (
 	"db_backup_go/common"
-	"db_backup_go/modules"
 	"db_backup_go/modules/clear"
 	"db_backup_go/modules/database"
 	"db_backup_go/modules/run"
@@ -79,7 +78,7 @@ func (fi fileInfo) Controller() error {
 	wg.Wait()
 
 	//按天保留最新7份备份，删除之前的备份
-	sshSocket := modules.NewSshSocket(conf.REMOTE_HOST, conf.REMOTE_PORT, conf.REMOTE_USER, conf.REMOTE_PASSWORD)
+	sshSocket := common.NewSshSocket(conf.REMOTE_HOST, conf.REMOTE_PORT, conf.REMOTE_USER, conf.REMOTE_PASSWORD)
 
 	rmFile := clear.NewBackupClear(conf.SAVE_DAY, *sshSocket)
 	rmFile.ClearLocal(conf.BACKUP_SAVE_PATH)
