+++ main.go
@@ -10,7 +10,7 @@ import (
 
 //程序主入口
 func main() {
-	version := "1.1.7"
+	version := "1.1.8"
 	printVersion := flag.Bool("version", false, "[--version]")
 	flag.Parse()
 	if *printVersion {
